+++ src/Texture.cpp
@@ -68,11 +68,13 @@ namespace tgui
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-#if TGUI_HAS_RENDERER_BACKEND_SFML_GRAPHICS
+#if TGUI_HAS_RENDERER_BACKEND_SFML_GRAPHICS && !defined(TGUI_REMOVE_DEPRECATED_CODE)
+TGUI_IGNORE_DEPRECATED_WARNINGS_START
     Texture::Texture(const sf::Texture& texture, const UIntRect& partRect, const UIntRect& middlePart)
     {
         load(texture, partRect, middlePart);
     }
+TGUI_IGNORE_DEPRECATED_WARNINGS_END
 #endif
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -210,7 +212,7 @@ namespace tgui
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-#if TGUI_HAS_RENDERER_BACKEND_SFML_GRAPHICS
+#if TGUI_HAS_RENDERER_BACKEND_SFML_GRAPHICS && !defined(TGUI_REMOVE_DEPRECATED_CODE)
     void Texture::load(const sf::Texture& texture, const UIntRect& partRect, const UIntRect& middleRect)
     {
         if (getData() && (m_destructCallback != nullptr))
@@ -341,6 +343,28 @@ namespace tgui
 #endif
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+    void Texture::setMiddleRect(const UIntRect& middleRect)
+    {
+        if (middleRect == UIntRect{})
+            m_middleRect = {0, 0, m_partRect.width, m_partRect.height};
+        else
+        {
+            m_middleRect = middleRect;
+
+            // If the middle rect was only partially provided then we need to calculate the width and height ourselves
+            if (((middleRect.left > 0) || (middleRect.top > 0)) && (middleRect.width == 0) && (middleRect.height == 0))
+            {
+                if (m_partRect.width > 2 * middleRect.left)
+                    m_middleRect.width = m_partRect.width - (2 * middleRect.left);
+
+                if (m_partRect.height > 2 * middleRect.top)
+                    m_middleRect.height = m_partRect.height - (2 * middleRect.top);
+            }
+        }
+    }
+
+    /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
     UIntRect Texture::getMiddleRect() const
     {
         return m_middleRect;
@@ -462,22 +486,7 @@ namespace tgui
         else
             m_partRect = partRect;
 
-        if (middleRect == UIntRect{})
-            m_middleRect = {0, 0, m_partRect.width, m_partRect.height};
-        else
-        {
-            m_middleRect = middleRect;
-
-            // If the middle rect was only partially provided then we need to calculate the width and height ourselves
-            if (((middleRect.left > 0) || (middleRect.top > 0)) && (middleRect.width == 0) && (middleRect.height == 0))
-            {
-                if (m_partRect.width > 2 * middleRect.left)
-                    m_middleRect.width = m_partRect.width - (2 * middleRect.left);
-
-                if (m_partRect.height > 2 * middleRect.top)
-                    m_middleRect.height = m_partRect.height - (2 * middleRect.top);
-            }
-        }
+        setMiddleRect(middleRect);
     }
 
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////